Draft of new shadowlru shadow slabs.
Guaranteed to be broken for the moment, but compiles and captures
the rough idea.

All policies need to be double checked to fix up return values
on proc() now as a result of the change in the meaning
(to return stack distance in bytes rather than cache size delta).

+++ fifo.cpp
@@ -18,7 +18,7 @@ fifo::~fifo () {
 // checks the hashmap for membership, if the key is found
 // returns a hit, otherwise the key is added to the hash
 // and to the FIFO queue and returns a miss.
-int64_t fifo::proc(const request *r, bool warmup) {
+size_t fifo::proc(const request *r, bool warmup) {
   if (!warmup)
     ++accesses;
