+++ lru.h
@@ -13,7 +13,7 @@ class lru : public policy {
   public:
     lru(uint64_t size);
    ~lru();
-    int64_t proc(const request *r, bool warmup);
+    size_t proc(const request *r, bool warmup);
     uint32_t get_size();
 
     size_t get_hits();
